+++ src/ui/MiniShopLayer.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "MiniLayer.hpp"
+#include <Geode/binding/GJShopLayer.hpp>
+
+using namespace geode::prelude;
+
+class MiniShopLayer : public MiniLayer<GJShopLayer> {
+protected:
+    bool init(ShopType type);
+public:
+    static MiniShopLayer* create(ShopType type);
+};
+
+class JamMarketLayer : public CCLayer {};
+
+class MiniJamShopLayer : public MiniLayer<JamMarketLayer> {
+protected:
+    bool init(JamMarketLayer* layer);
+public:
+    static MiniJamShopLayer* create(JamMarketLayer* layer);
+};